web_server/requirements.txt
annotated-types==0.6.0
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.0
google_images_download==2.8.0
googleapis-common-protos==1.63.0
googlesearch-python==1.2.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httplib2==0.22.0
idna==3.7
outcome==1.3.0.post0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
PySocks==1.7.1
requests==2.31.0
rsa==4.9
selenium==4.19.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wsproto==1.2.0